+++ packages/base-types/src/node/utils/trace.ts
@@ -29,4 +29,5 @@ export interface BaseTraceFrame<Payload = any, TracePath extends BaseTraceFrameP
   paths?: TracePath[];
   payload: Payload;
   defaultPath?: number;
+  stop?: boolean;
 }