+++ packages/google-dfes-types/src/utils/index.ts
@@ -1,2 +1,3 @@
+export * as mappings from './mappings';
 export * as node from './node';
 export * as step from './step';