+++ packages/google-types/src/constants/index.ts
@@ -1,5 +1,6 @@
 export * from './configFile';
 export * from './intents';
 export * from './locales';
+export * from './mappings';
 export * from './slots';
 export * from './voices';