+++ packages/google-dfes-types/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0](https://github.com/voiceflow/libs/compare/@voiceflow/google-dfes-types@2.10.32...@voiceflow/google-dfes-types@2.11.0) (2022-06-21)
+
+
+### Features
+
+* move mappings from google service to libs (VF-000) ([#309](https://github.com/voiceflow/libs/issues/309)) ([0181459](https://github.com/voiceflow/libs/commit/018145919fda0c33e2dd18fa68e016c292e4b4b4))
+
+
+
+
+
+## [2.10.32](https://github.com/voiceflow/libs/compare/@voiceflow/google-dfes-types@2.10.31...@voiceflow/google-dfes-types@2.10.32) (2022-06-21)
+
+**Note:** Version bump only for package @voiceflow/google-dfes-types
+
+
+
+
+
+## [2.10.31](https://github.com/voiceflow/libs/compare/@voiceflow/google-dfes-types@2.10.30...@voiceflow/google-dfes-types@2.10.31) (2022-06-17)
+
+**Note:** Version bump only for package @voiceflow/google-dfes-types
+
+
+
+
+
 ## [2.10.30](https://github.com/voiceflow/libs/compare/@voiceflow/google-dfes-types@2.10.29...@voiceflow/google-dfes-types@2.10.30) (2022-06-15)
 
 **Note:** Version bump only for package @voiceflow/google-dfes-types